Extract triggers from automation integration (#609)
